api/feat GET /health
Health endpoint which checks the clients the SentioVoice API needs to work.

If the API has been able to instantiate all the clients (i.e.; no configuration is missing),
then a 500 error is thrown.

In both cases (200 and 500 responses), the API returns a JSON object with a key corresponding each client,
and a boolean value to point out the status of the client.